refactor: move Fieldset component
