+++ packages/big-design/src/components/Fieldset/styled.tsx
@@ -1,7 +1,7 @@
 import { theme as defaultTheme } from '@bigcommerce/big-design-theme';
 import styled, { DefaultTheme, StyledComponent } from 'styled-components';
 
-import { StyleableH3, StyleableSmall } from '../../Typography/private';
+import { StyleableH3, StyleableSmall } from '../Typography/private';
 
 export const StyledFieldset = styled.fieldset`
   border: none;
